fix: predicates helpers
